fix(grpc): leaking spans when using gRPC future interface
We've tracked down some issues related to broken span hierarchy down to
the way that the gRPC instrumentation deals with entering spans, which
causes the grpc span to leak outside it's call when using the .future()
interface. This PR changes the instrumentation to explicitly only enter
the span when calling the continuations instead of just entering
immediately.

Signed-off-by: Sam Schlegel <sam.s@discordapp.com>